# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteAccountResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteAccountResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileOut"
          },
          "saved_story_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Saved Story Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "profile"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "NotificationOut": {
        "description": "\u00a79.4 notification history \u2014 deep link is `story_id` (None means the\nhome feed, e.g. DAILY_BRIEFING or a suppressed/retracted target).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "enum": [
              "PENDING",
              "SENT",
              "FAILED",
              "SUPPRESSED"
            ],
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Id"
          },
          "suppressed_reason": {
            "anyOf": [
              {
                "enum": [
                  "QUIET_HOURS",
                  "DAILY_CAP"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suppressed Reason"
          },
          "type": {
            "enum": [
              "DAILY_BRIEFING",
              "TOPIC_ALERT",
              "BREAKING_ALERT"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "story_id",
          "status",
          "sent_at",
          "created_at"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
...
      "PreferencesUpdate": {
        "properties": {
          "breaking_alerts_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breaking Alerts Enabled"
          },
          "daily_briefing_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Briefing Enabled"
          },
          "home_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home City"
          },
          "home_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home State"
          },
          "language": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "te"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "max_alerts_per_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Alerts Per Day"
          },
          "notification_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Mode"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "residence_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence Country"
          },
          "residence_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence Region"
          },
          "topic_slugs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Slugs"
          }
        },
        "title": "PreferencesUpdate",
        "type": "object"
      },
      "ProfileOut": {
        "properties": {
          "breaking_alerts_enabled": {
            "default": true,
            "title": "Breaking Alerts Enabled",
            "type": "boolean"
          },
          "daily_briefing_enabled": {
            "default": true,
            "title": "Daily Briefing Enabled",
            "type": "boolean"
          },
          "home_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home City"
          },
          "home_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home State"
          },
          "language": {
            "default": "en",
            "enum": [
              "en",
              "te"
            ],
            "title": "Language",
            "type": "string"
          },
          "max_alerts_per_day": {
            "default": 5,
            "title": "Max Alerts Per Day",
            "type": "integer"
          },
          "notification_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Mode"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "residence_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence Country"
          },
          "residence_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence Region"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/TopicOut"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "title": "ProfileOut",
        "type": "object"
      },
      "PushTokenCreate": {
        "properties": {
          "platform": {
            "enum": [
              "ios",
              "android",
              "web"
            ],
            "title": "Platform",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "platform"
        ],
        "title": "PushTokenCreate",
        "type": "object"
      },
      "PushTokenResponse": {
        "properties": {
          "registered": {
            "title": "Registered",
            "type": "boolean"
          }
        },
        "required": [
          "registered"
        ],
        "title": "PushTokenResponse",
        "type": "object"
      },
...
      "SavedStoryResponse": {
        "properties": {
          "saved": {
            "title": "Saved",
            "type": "boolean"
          },
          "story_id": {
            "format": "uuid",
            "title": "Story Id",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "saved"
        ],
        "title": "SavedStoryResponse",
        "type": "object"
      },
...
      "TopicOut": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "TopicOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Telugu Global API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "get_me_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/account": {
      "delete": {
        "operationId": "delete_account_v1_me_account_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Account",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "operationId": "list_notifications_v1_me_notifications_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "title": "Response List Notifications V1 Me Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notifications",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/preferences": {
      "patch": {
        "operationId": "update_preferences_v1_me_preferences_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preferences",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/push-tokens": {
      "post": {
        "operationId": "register_push_token_v1_me_push_tokens_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Push Token",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/saved/{story_id}": {
      "delete": {
        "operationId": "unsave_story_v1_me_saved__story_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Story Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedStoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsave Story",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "save_story_v1_me_saved__story_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Story Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedStoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Story",
        "tags": [
          "me"
        ]
      }
    },
...
  }
}